+++ DLC/LEKMOD_V28.2/nqmod-vs2008-master/CvGameCoreDLL_Expansion2/CvPlot.cpp
@@ -3527,7 +3527,7 @@ int CvPlot::GetAircraftCapacity() const
 
 		// units
 		const vector<const CvUnit*> units = getAllUnitsConst();
-		for (int i = 0; i < units.size(); ++i)
+		for (int i = 0; i < (int)units.size(); ++i)
 		{
 			const CvUnit* unit = units[i];
 			if (unit)
